monitoring/requirements.txt
# Kafka client for Python
kafka-python==2.0.2

# Elasticsearch client (compatible with Elasticsearch 8.x)
elasticsearch>=8.0.0,<9.0.0

# Logging and utilities
python-dotenv==1.0.0
